requirements.txt
# Core Trading Libraries
ccxt>=4.2.0
python-dotenv>=1.0.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
pandas-ta>=0.3.14b

# Machine Learning
scikit-learn>=1.3.0
joblib>=1.3.0

# Natural Language Processing
textblob>=0.17.1
requests>=2.31.0

# Logging & Database
sqlalchemy>=2.0.0

# Visualization (for backtesting)
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration
pyyaml>=6.0.0

# Testing
pytest>=7.4.0
pytest-mock>=3.11.0

# Phase 2: Advanced Features
# Deep Learning & NLP
transformers>=4.30.0  # For FinBERT sentiment analysis
torch>=2.0.0  # For LSTM neural networks
sentencepiece>=0.1.99  # Required by transformers

# Telegram Bot
python-telegram-bot>=20.0  # For notifications and remote control

# Real-time Data
websockets>=11.0  # For WebSocket streaming (optional)

# Additional ML
xgboost>=2.0.0  # Gradient boosting (optional)
lightgbm>=4.0.0  # Light gradient boosting (optional)

# Web Dashboard
flask>=3.0.0
flask-socketio>=5.3.0
python-socketio>=5.10.0
gevent>=23.9.0
